Added movegen test.
The movegen tests the move generation code by seeing if it generates all moves
used in a full game. To test this, a sample of many games (thousands) from
(http://theweekinchess.com/twic) is used.

This catches bugs where the move generation doesn't find a valid move in a given position,
as well as when it updates board state incorrectly, such as with en passant or castling.

It doesn't catch extra moves that are generated that may not be legal.