Removed invalid test data from Movegen test
There were a number of problems with the test data. First, PGN format normally looks like this:

[tags]

move1 move2 ... moveX 0-1

However, some of the games in the test files were like this:

[tags]

0-1

They had no movelist. pgn-extract passed this game on as a "0-1" in the output file, which makes the test fail because
it isn't a list of moves. To fix this, the output of pgn-extract is now run through filtering to remove these lines. This can be seen in the
script file convert-pgn-to-san.

The second problem was that some of the files had extra empty lines between games, or at the end. This should be permissible input, but the code
previously assumed that there were never multiple blank lines in a row. This has been fixed so blank lines are completely ignored, and any non-blank line
is assumed to be a move-list.

The third problem was with the dataset found in twic1208.pgn. It contained at least 1 game with an illegal move. It appeared that it was a blitz match
where an illegal move was made, followed immediately by the game result being a win for the opponent, that is, a forfeit. The test data should only contain legal games, so
for the time being, twic1208.pgn was removed as input to the test.