Fixed castling state bug triggered by rook captures
When a rook is captured, the ability to castle on a certain side may be lost. The code that detected which side,
of both board (queenside, kingside) and turn (white, black) was invalid. For example, if white captured a black rook on the a file,
white would lose the ability to castle queenside, because a rook captured occurred on the file of the queenside rook. The assumption made was that
the rook capture occurred to the side moving, although that is wrong 100% of the time. The code was changed to check for the exact rook squares in question, regardless of turn,
and simplified somewhat.